src/tchapi/requirements.txt
adal==1.2.7
annotated-types==0.7.0
applicationinsights==0.11.10
arch==5.6.0
argcomplete==3.4.0
asttokens==2.4.1
attrs==24.1.0
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==4.7.0
azure-graphrbac==0.61.1
azure-identity==1.17.1
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==26.0.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.2.1
azure-storage-blob==12.19.0
azure-storage-queue==12.11.0
azureml-automl-core==1.57.0
azureml-automl-runtime==1.57.0
azureml-core==1.57.0.post3
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.2
azureml-dataset-runtime==1.57.0
azureml-defaults==1.57.0
azureml-inference-server-http==1.2.2
azureml-interpret==1.57.0
azureml-mlflow==1.57.0
azureml-pipeline-core==1.57.0
azureml-responsibleai==1.57.0
azureml-telemetry==1.57.0
azureml-train-automl==1.57.0
azureml-train-automl-client==1.57.0
azureml-train-automl-runtime==1.57.0
azureml-train-core==1.57.0
azureml-train-restclients-hyperdrive==1.57.0
azureml-training-tabular==1.57.0.post1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.0
blinker==1.8.2
bokeh==2.4.3
boto3==1.20.19
botocore==1.23.19
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cmdstanpy==1.2.4
coloredlogs==15.0.1
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.1
convertdate==2.4.0
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.11
dask==2023.2.0
databricks-cli==0.18.0
dataclasses==0.6
debugpy==1.8.2
decorator==5.1.1
dice-ml==0.11
dill==0.3.8
distributed==2023.2.0
distro==1.9.0
docker==7.1.0
dotnetcore2==3.1.23
econml==0.15.1
entrypoints==0.4
ephem==4.1.5
erroranalysis==0.5.4
exceptiongroup==1.2.2
executing==2.0.1
fairlearn==0.7.0
fire==0.6.0
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
fusepy==3.0.1
gensim==4.3.2

gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.32.0
googleapis-common-protos==1.63.2
gunicorn==22.0.0
holidays==0.57
humanfriendly==10.0
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
inference-schema==1.7.2
interpret-core==0.5.0
interpret_community==0.31.0
ipykernel==6.29.5
ipython==8.18.1
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.2.0
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras2onnx==1.6.0
kiwisolver==1.4.5
knack==0.11.0
lightgbm==3.2.1
llvmlite==0.39.1
locket==1.0.0
LunarCalendar==0.0.9
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
ml_wrappers==0.5.5
mlflow-skinny==2.9.2
mltable==1.6.1
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
msgpack==1.0.8
msrest==0.7.1
msrestazure==0.6.4.post1
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
networkx==2.5
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.16.1
onnxconverter-common==1.13.0
onnxmltools==1.11.2
onnxruntime==1.17.3
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
packaging==23.2
pandas==1.3.5
paramiko==3.4.0
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
pmdarima==1.8.5
portalocker==2.10.1
prompt_toolkit==3.0.47
property-cached==1.6.4
prophet==1.1.4
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==14.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PyMeeus==0.5.12
PyNaCl==1.5.0
pyogrio==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.1.0
raiutils==0.4.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
responsibleai==0.36.0
rpds-py==0.19.1
rsa==4.9
s3transfer==0.5.2
scikit-learn==1.5.1
scipy==1.10.1
SecretStorage==3.3.3
semver==2.13.0
shap==0.43.0
shapely==2.0.6
six==1.16.0
skl2onnx==1.15.0
sklearn-pandas==1.7.0
slicer==0.0.7
smart-open==6.4.0
smmap==5.0.1
sortedcontainers==2.4.0
sparse==0.15.4
sqlparse==0.5.1
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.13.5
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
termcolor==2.4.0
threadpoolctl==3.5.0
toolz==0.12.1
torch==1.13.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.19
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==1.5.2
zict==3.0.0
zipp==3.19.2
